# backend/docs/swagger.yaml
basePath: /
definitions:
...
  controllers.PresignDownloadRequest:
    properties:
      bucket:
        type: string
      expire:
        description: 过期时间（秒），默认3600
        type: integer
      key:
        type: string
    required:
    - bucket
    - key
    type: object
  controllers.PresignDownloadResponse:
    properties:
      expires_in:
        type: integer
      url:
        type: string
    type: object
  controllers.PresignUploadRequest:
    properties:
      bucket:
        type: string
      expire:
        description: 过期时间（秒），默认3600
        type: integer
      key:
        type: string
    required:
    - bucket
    - key
    type: object
  controllers.PresignUploadResponse:
    properties:
      expires_in:
        type: integer
      url:
        type: string
    type: object
...
  controllers.RustFSBucketInfo:
    properties:
      CreationDate:
        type: string
      Name:
        type: string
    type: object
  controllers.RustFSObjectInfo:
    properties:
      ETag:
        type: string
      Key:
        type: string
      LastModified:
        type: string
      Size:
        type: integer
      StorageClass:
        type: string
    type: object
  controllers.TemplateCreateRequest:
    properties:
      height:
        type: integer
      name:
        type: string
      offset_x:
        type: number
      offset_y:
        type: number
      scale:
        type: number
      url:
        type: string
      user_id:
        type: integer
      width:
        type: integer
    required:
    - name
    - url
    type: object
  controllers.TemplateUpdateRequest:
    properties:
      height:
        type: integer
      name:
        type: string
      offset_x:
        type: number
      offset_y:
        type: number
      scale:
        type: number
      url:
        type: string
      width:
        type: integer
    type: object
...
  models.Template:
    properties:
      created_at:
        type: string
      height:
        description: 素材高度
        type: integer
      id:
        type: integer
      name:
        description: 模版名称
        type: string
      offset_x:
        description: X轴偏移
        type: number
      offset_y:
        description: Y轴偏移
        type: number
      scale:
        description: 缩放比例
        type: number
      updated_at:
        type: string
      url:
        description: 模版URL
        type: string
      user_id:
        description: 用户ID
        type: integer
      width:
        description: 素材宽度
        type: integer
    type: object
...
  utils.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: xhw-service 后台接口文档，业务接口统一返回 utils.Response，鉴权接口请先登录后在 Swagger Authorize
    中填写 Bearer Token。
  title: xhw-service API
  version: "1.0"
paths:
...
  /api/v1/rustfs/buckets:
    get:
      description: 获取 RustFS 中的全部存储桶。
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/controllers.RustFSBucketInfo'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 获取存储桶列表
      tags:
      - 对象存储
  /api/v1/rustfs/buckets/{bucket}:
    post:
      description: 创建指定名称的存储桶。
      parameters:
      - description: 存储桶名称
        in: path
        name: bucket
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 创建存储桶
      tags:
      - 对象存储
  /api/v1/rustfs/buckets/{bucket}/objects:
    get:
      description: 获取指定存储桶下的对象列表。
      parameters:
      - description: 存储桶名称
        in: path
        name: bucket
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/controllers.RustFSObjectInfo'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 获取对象列表
      tags:
      - 对象存储
  /api/v1/rustfs/buckets/{bucket}/objects/{key}:
    delete:
      description: 删除指定存储桶下的对象，key 支持携带多级路径。
      parameters:
      - description: 存储桶名称
        in: path
        name: bucket
        required: true
        type: string
      - description: 对象 Key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 删除对象
      tags:
      - 对象存储
  /api/v1/rustfs/presign/download:
    post:
      consumes:
      - application/json
      description: 为指定 bucket/key 生成 GET 下载预签名 URL，expire 为过期秒数，默认 3600。
      parameters:
      - description: 下载预签名参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.PresignDownloadRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/controllers.PresignDownloadResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 获取下载预签名 URL
      tags:
      - 对象存储
  /api/v1/rustfs/presign/upload:
    post:
      consumes:
      - application/json
      description: 为指定 bucket/key 生成 PUT 上传预签名 URL，expire 为过期秒数，默认 3600。
      parameters:
      - description: 上传预签名参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.PresignUploadRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/controllers.PresignUploadResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 获取上传预签名 URL
      tags:
      - 对象存储
  /api/v1/templates:
    get:
      description: 管理员可按 user_id 查询任意用户模版，普通用户仅能查询自己的模版。
      parameters:
      - description: 用户ID，管理员可选
        in: query
        name: user_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Template'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 获取模版列表
      tags:
      - 模版管理
    post:
      consumes:
      - application/json
      description: 创建模版；普通用户提交的 user_id 会被强制覆盖为当前登录用户，管理员可指定 user_id。
      parameters:
      - description: 模版信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.TemplateCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Template'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 创建模版
      tags:
      - 模版管理
  /api/v1/templates/{id}:
    delete:
      description: 删除模版，普通用户只能删除自己的模版。
      parameters:
      - description: 模版ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 删除模版
      tags:
      - 模版管理
    get:
      description: 管理员可查看任意模版，普通用户只能查看自己的模版。
      parameters:
      - description: 模版ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Template'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 获取模版详情
      tags:
      - 模版管理
    put:
      consumes:
      - application/json
      description: 更新模版信息，普通用户只能更新自己的模版，user_id 不允许修改。
      parameters:
      - description: 模版ID
        in: path
        name: id
        required: true
        type: integer
      - description: 模版更新内容
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.TemplateUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Template'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 更新模版
      tags:
      - 模版管理
...
securityDefinitions:
  BearerAuth:
    description: JWT Authorization header，格式为 Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"